requirements.txt
auto_mix_prep==0.2.0
pythonnet==3.0.0a1
clr-loader==0.1.7
numpy
pandas
python-dotenv==0.19.2
pytest
black
papermill
jupytext
requests
